LayoutTests/imported/w3c/web-platform-tests/html/syntax/speculative-parsing/generated/page-load/picture-source-no-img.tentative-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: image/webp,image/avif,image/jxl,image/heic,image/heic-sequence,video/*;q=0.8,imag...


FAIL Speculative parsing, page load: picture-source-no-img Unhandled rejection: assert_equals: speculative case incorrectly fetched expected "" but got "param-encodingcheck: %C4%9E\r\nAccept: image/webp,image/avif,image/jxl,image/heic,image/heic-sequence,video/*;q=0.8,image/png,image/svg+xml,image/*;q=0.8,*/*;q=0.5\r\nReferer: http://web-platform.test:8800/html/syntax/speculative-parsing/generated/page-load/resources/picture-source-no-img-framed.sub.html?uuid=2a7eab0a-766e-4012-baef-12bd9a619d91"
